+++ Source/NotifyIconWpf/Interop/WindowClass.cs
@@ -7,7 +7,7 @@ namespace Hardcodet.Wpf.TaskbarNotification.Interop
   /// Callback delegate which is used by the Windows API to
   /// submit window messages.
   /// </summary>
-  public delegate long WindowProcedureHandler(IntPtr hwnd, uint uMsg, uint wparam, uint lparam);
+  public delegate IntPtr WindowProcedureHandler(IntPtr hwnd, uint uMsg, IntPtr wparam, IntPtr lparam);
 
 
   /// <summary>
